+++ src/grid.hpp
@@ -2,19 +2,20 @@
 
 #include <compare>
 #include <cstddef>
+#include <cstdint>
 #include <functional>
 
 using ThreadCount = int;
 
 struct GridSquare {
-    long long x, y;
+    int64_t x, y;
     friend auto operator<=>(GridSquare, GridSquare) = default;
 };
 
 template <>
 struct std::hash<GridSquare> {
     size_t operator()(GridSquare square) const noexcept {
-        auto mix = [](unsigned long long x) {
+        auto mix = [](uint64_t x) {
             x += 0x9e3779b97f4a7c15;
             x = (x ^ (x >> 30)) * 0xbf58476d1ce4e5b9;
             x = (x ^ (x >> 27)) * 0x94d049bb133111eb;
